% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_tails.R
\name{mutate_tails}
\alias{mutate_tails}
\title{Mutating existing normal deviation to heavy-tailed ones}
\usage{
mutate_tails(
  x,
  dist = c("student-t", "pareto"),
  nout = length(x),
  bootstrap = FALSE,
  args = list()
)
}
\arguments{
\item{x}{vector of random deviates, must be positive and centered around one.}

\item{dist}{what type of distribution to convert to. Options are `student-t`
and `pareto`.}

\item{nout}{length of output vector. Defaults to length(x). This is only
used when `bootstrap = TRUE`.}

\item{bootstrap}{logical, whether to bootstrap samples? If `TRUE`, then
number of samples is determined by `nout`. Bootstraps are with replacement.}

\item{args}{list of arguments to pass to the quantile function.
If `dist = student-t`, then a `df` argument is required, while if
`dist = pareto` then a shape argument is required.}
}
\description{
Random generation for a mixture of normal distributions, both centered
around zero and bias corrected, as well as exponentiated.
}
\details{
This function takes a series of normally-distributed deviations
and "stretches" them into a heavy-tailed distribution. It uses `pnorm()` to
transform to a uniform distribution and then a quantile function to
transform to a distribution of type `dist`.
}
\examples{
recdevs <- exp(rnorm(60, -0.5 * 0.1^2, 0.1)) # bias-corrected
recdevs_st <- mutate_tails(recdevs, dist = "student-t", args = list(df = 0.5))
plot_tails(recdevs)
plot_tails(recdevs_st)

recdevs_p <- mutate_tails(recdevs, dist = "pareto", args = list(shape = 1.005))
plot_tails(recdevs_p)

}
