% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tails.R
\name{plot_tails}
\alias{plot_tails}
\title{Plot heavy-tailed deviations}
\usage{
plot_tails(vec)
}
\arguments{
\item{vec}{Vector of random deviates.}
}
\value{
A scatter plot showing the deviation values in the vector.
Sample mean is shown by green line, and the marginal distribution is
shown by the rug plot on the right side of the plot.
}
\description{
A simple graphical representation of random heavy-tailed deviations created
with `rpareto_tails()`, `rst_tails()`, `rnorm_tails()`, or any other
multiplicative random deviates.
}
\details{
If `vec` is the output of `rnorm_tails()` where sporadic values
have been included, these values are marked in red.
}
\examples{
rpt <- rpareto_tails(50, 3)
plot_tails(rpt)

}
