% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgev_tails.R
\name{rgev_tails}
\alias{rgev_tails}
\title{Random generation of generalized extreme value deviations}
\usage{
rgev_tails(
  n,
  scale = 1,
  shape = 0,
  bias_correct = TRUE,
  ac = 0,
  log = FALSE,
  seed = NA
)
}
\arguments{
\item{n}{sample size.}

\item{scale}{scale parameter. Must be > 0.}

\item{shape}{shape parameter. Default is shape = 2.}

\item{bias_correct}{logical. Should we bias correct using the sample mean?}

\item{ac}{auto-correlation value, between -1 and 1. If `ac != 0`
autocorrelation is incorporated in the vector using an AR(\emph{1}) process.}

\item{log}{logical. Whether to return the log-transformed distribution.}

\item{seed}{seed. Numeric for `set.seed()`. Defaults to NA where no seed is
set.}

\item{loc}{location parameter.}
}
\description{
Random generation for the generalized extreme value (GEV) distribution,
centered around a mean of one, with parameter `shape`.
}
\examples{
rgev <- rgev_tails(50, 0, 1, bias_correct = TRUE)
plot_tails(rgev)

}
