% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_tails.R
\name{rnorm_tails}
\alias{rnorm_tails}
\title{Random generation of normal mixture deviations}
\usage{
rnorm_tails(n, sigma = 0.1, highsigma = 2, rate = 1/38, replace = FALSE)
}
\arguments{
\item{n}{sample size, or a vector of random deviates to update if
`replace = TRUE`.}

\item{sigma}{variance of the underlying normal distribution, defaults
to 0.1. Ignored if `replace = TRUE`.}

\item{highsigma}{variance of the sporadic normal distribution, should be
higher than `sigma`, defaults to 1.5.}

\item{rate}{rate at which the wider sporadic normal distribution replaces
the underlying distribution. This is determined by a binomial process
using `rbinom(1, n, rate)`.}

\item{replace}{logical. Whether to replace a vector of deviates using
sporadic values or create it from scratch. If `TRUE`, then
`n` needs to be a vector of positive deviates centered around 1.}
}
\value{
a vector of random deviates of length `n`. If there were sporadic
deviates where the `highsigma` distribution replaced the underlying one,
then the positions of these events are stored in the `ht` attribute of the
vector.
}
\description{
Random generation for a mixture of normal distributions, both centered
around zero and bias corrected, as well as exponentiated.
}
\details{
Both normal distributions are bias corrected independently so that
their means in log-space equal to 1 (\eqn{-\sigma^2/2}).
}
\examples{
rnt <- rnorm_tails(500, sigma = 0.2, highsigma = 1.5, rate = 1/19)
plot_tails(rnt)

}
