% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpareto_tails.R
\name{rpareto_tails}
\alias{rpareto_tails}
\title{Random generation of Pareto deviations}
\usage{
rpareto_tails(n, shape = 2)
}
\arguments{
\item{n}{sample size.}

\item{shape}{shape parameter. Default is shape = 2.}
}
\description{
Random generation for the Pareto distribution, centered around a mean
of one, with parameter `shape`.
}
\examples{
rpt <- rpareto_tails(50, 3)
plot_tails(rpt)

}
