% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst_tails.R
\name{rst_tails}
\alias{rst_tails}
\title{Random generation of Student-t deviations}
\usage{
rst_tails(n, df = 10, bias.correct = FALSE)
}
\arguments{
\item{n}{sample size.}

\item{df}{degrees of freedom (> 0). Default is df = 10. df = -Inf is
allowed if bias.correct = FALSE, where it becomes equivalent to the standard
normal distribution.}

\item{bias.correct}{Logical. Should we bias correct the mean using
the log-normal distribution correction factor (\eqn{-\sigma^2/2}).}
}
\value{
a vector of random deviates of length `n`.
}
\description{
Random generation for the Student-t distribution, centered around zero,
exponentiated, with `df` degrees of freedom.
}
\examples{
rst <- rst_tails(50, 3)
plot_tails(rst)

}
