% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tails.R
\name{plot_tails}
\alias{plot_tails}
\title{Plot heavy-tailed deviations}
\usage{
plot_tails(vec)
}
\arguments{
\item{vec}{Vector of random deviates.}
}
\description{
A simple graphical representation of random heavy-tailed deviations created
with `rpareto_tail()`...
}
\examples{
rpt <- rpareto_tails(50, 3)
plot_tails(rpt)

}
